refactor: admin 실행 경계를 context 단위로 고정
Constraint: 기존 admin endpoint와 JSON request/response field를 변경하지 않아야 함
Rejected: 별도 file context 도입 | 현재 presigned-url 유스케이스는 carousel/banner promotion asset 흐름에 묶여 있음
Rejected: root admin api/facade/application service 유지 | Kotlin migration 이후 package split diff가 중복됨
Confidence: high
Scope-risk: moderate
Directive: admin DTO는 domain type을 import하지 말고 context application service 안에서 primitive/result로 변환할 것
Tested: ./gradlew :admin:compileJava :admin:compileKotlin :admin:test :test --tests com.beat.architecture.PromotionBoundaryTest --no-daemon
Tested: ./gradlew check --no-daemon
