--- admin/src/main/java/com/beat/admin/application/dto/request/AdminCarouselNumber.java
@@ -1,15 +1,0 @@
-package com.beat.admin.application.dto.request;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(name = "CarouselNumber")
-public enum AdminCarouselNumber {
-	ONE,
-	TWO,
-	THREE,
-	FOUR,
-	FIVE,
-	SIX,
-	SEVEN,
-	;
-}
