--- admin/src/main/java/com/beat/admin/application/dto/response/CarouselPresignedUrlFindAllResponse.java
@@ -1,17 +1,0 @@
-package com.beat.admin.application.dto.response;
-
-import java.util.Map;
-
-import com.beat.contracts.storage.CarouselPresignedUrls;
-
-public record CarouselPresignedUrlFindAllResponse(
-	Map<String, String> carouselPresignedUrls
-) {
-	private static CarouselPresignedUrlFindAllResponse of(Map<String, String> carouselPresignedUrls) {
-		return new CarouselPresignedUrlFindAllResponse(carouselPresignedUrls);
-	}
-
-	public static CarouselPresignedUrlFindAllResponse from(CarouselPresignedUrls response) {
-		return CarouselPresignedUrlFindAllResponse.of(response.carouselPresignedUrls());
-	}
-}
