+++ admin/src/main/java/com/beat/admin/promotion/application/dto/response/BannerPresignedUrlFindResponse.java
@@ -1,0 +1,15 @@
+package com.beat.admin.promotion.application.dto.response;
+
+import com.beat.contracts.storage.BannerPresignedUrl;
+
+public record BannerPresignedUrlFindResponse(
+	String bannerPresignedUrl
+) {
+	private static BannerPresignedUrlFindResponse of(String bannerPresignedUrl) {
+		return new BannerPresignedUrlFindResponse(bannerPresignedUrl);
+	}
+
+	public static BannerPresignedUrlFindResponse from(BannerPresignedUrl response) {
+		return BannerPresignedUrlFindResponse.of(response.bannerPresignedUrl());
+	}
+}
