+++ admin/src/main/java/com/beat/admin/promotion/application/dto/response/CarouselPresignedUrlFindAllResponse.java
@@ -1,0 +1,17 @@
+package com.beat.admin.promotion.application.dto.response;
+
+import java.util.Map;
+
+import com.beat.contracts.storage.CarouselPresignedUrls;
+
+public record CarouselPresignedUrlFindAllResponse(
+	Map<String, String> carouselPresignedUrls
+) {
+	private static CarouselPresignedUrlFindAllResponse of(Map<String, String> carouselPresignedUrls) {
+		return new CarouselPresignedUrlFindAllResponse(carouselPresignedUrls);
+	}
+
+	public static CarouselPresignedUrlFindAllResponse from(CarouselPresignedUrls response) {
+		return CarouselPresignedUrlFindAllResponse.of(response.carouselPresignedUrls());
+	}
+}
