+++ admin/src/test/kotlin/com/beat/admin/AdminArchitectureGuardTest.kt
@@ -8,7 +8,7 @@ import java.nio.file.Path
 
 class AdminArchitectureGuardTest {
 
-    private val rootProjectDependencyPattern = Regex("""project\(\s*":\"\s*\)""")
+    private val rootProjectDependencyPattern = Regex("""project\(\s*":"\s*\)""")
 
     @Test
     fun `admin build file must not depend on root project`() {
@@ -73,12 +73,19 @@ class AdminArchitectureGuardTest {
     }
 
     @Test
-    fun `admin facade must not own transaction repository or raw domain model dependencies`() {
-        val source = Files.readString(Path.of("src/main/java/com/beat/admin/facade/AdminFacade.java"))
+    fun `admin facades must not own transaction repository or raw domain model dependencies`() {
+        val violations = findSourceViolationsInMatchingPaths(
+            Path.of("src/main/java/com/beat/admin"),
+            pathSegment = "/facade/",
+            "@Transactional",
+            "com.beat.domain.",
+            "Repository",
+        )
 
-        assertFalse(source.contains("@Transactional"))
-        assertFalse(source.contains("com.beat.domain."))
-        assertFalse(source.contains("Repository"))
+        assertTrue(
+            violations.isEmpty(),
+            "Found forbidden facade dependencies:\n${violations.joinToString("\n")}"
+        )
     }
 
     @Test
@@ -89,24 +96,35 @@ class AdminArchitectureGuardTest {
     }
 
     @Test
-    fun `admin does not keep transitional adapter or controller packages`() {
-        val forbiddenPackages = listOf(
+    fun `admin does not keep transitional adapter controller or root layer files`() {
+        val forbiddenPaths = listOf(
             Path.of("src/main/java/com/beat/admin/adapter"),
             Path.of("src/main/java/com/beat/admin/controller"),
+            Path.of("src/main/java/com/beat/admin/api/AdminApi.java"),
+            Path.of("src/main/java/com/beat/admin/api/AdminController.java"),
+            Path.of("src/main/java/com/beat/admin/facade/AdminFacade.java"),
+            Path.of("src/main/java/com/beat/admin/application/service"),
         )
 
-        val violations = forbiddenPackages.filter(Files::exists)
+        val violations = forbiddenPaths.filter(Files::exists)
 
         assertTrue(
             violations.isEmpty(),
-            "admin transitional HTTP package should not remain after api package normalization: ${violations.joinToString(", ")}"
+            "admin transitional package/file should not remain after context package split: ${violations.joinToString(", ")}"
         )
     }
 
     @Test
+    fun `admin success code belongs to api response boundary`() {
+        assertTrue(Files.exists(Path.of("src/main/kotlin/com/beat/admin/api/response/AdminSuccessCode.kt")))
+        assertFalse(Files.exists(Path.of("src/main/java/com/beat/admin/exception/AdminSuccessCode.java")))
+    }
+
+    @Test
     fun `admin dto layer must not import domain types`() {
-        val violations = findForbiddenImportsUnder(
-            Path.of("src/main/java/com/beat/admin/application/dto"),
+        val violations = findForbiddenImportsInMatchingPaths(
+            Path.of("src/main/java/com/beat/admin"),
+            pathSegment = "/dto/",
             "com.beat.domain."
         )
 
@@ -119,7 +137,7 @@ class AdminArchitectureGuardTest {
     @Test
     fun `admin application services do not return raw domain models`() {
         val violations = findMethodSignatureViolations(
-            Path.of("src/main/java/com/beat/admin/application"),
+            Path.of("src/main/java/com/beat/admin"),
             listOf("Promotion", "Users")
         )
 
@@ -151,7 +169,11 @@ class AdminArchitectureGuardTest {
         }
     }
 
-    private fun findForbiddenImportsUnder(root: Path, vararg forbiddenReferences: String): List<String> {
+    private fun findForbiddenImportsInMatchingPaths(
+        root: Path,
+        pathSegment: String,
+        vararg forbiddenReferences: String,
+    ): List<String> {
         if (!Files.exists(root)) {
             return emptyList()
         }
@@ -161,6 +183,7 @@ class AdminArchitectureGuardTest {
         return try {
             paths
                 .filter(Files::isRegularFile)
+                .filter { path -> path.toString().contains(pathSegment) }
                 .filter { path -> path.toString().endsWith(".java") || path.toString().endsWith(".kt") }
                 .toList()
                 .flatMap { path ->
@@ -173,6 +196,30 @@ class AdminArchitectureGuardTest {
                                 .map { pattern -> "$path: $pattern" }
                         }
                         .toList()
+                }
+        } finally {
+            paths.close()
+        }
+    }
+
+    private fun findSourceViolationsInMatchingPaths(
+        root: Path,
+        pathSegment: String,
+        vararg forbiddenReferences: String,
+    ): List<String> {
+        val paths = Files.walk(root)
+
+        return try {
+            paths
+                .filter(Files::isRegularFile)
+                .filter { path -> path.toString().contains(pathSegment) }
+                .filter { path -> path.toString().endsWith(".java") || path.toString().endsWith(".kt") }
+                .toList()
+                .flatMap { path ->
+                    val source = Files.readString(path)
+                    forbiddenReferences
+                        .filter(source::contains)
+                        .map { pattern -> "$path: $pattern" }
                 }
         } finally {
             paths.close()
