+++ src/test/java/com/beat/architecture/PromotionBoundaryTest.java
@@ -26,20 +26,20 @@ class PromotionBoundaryTest {
 	void executablePromotionFlowsUseDomainRepositoryContractInsteadOfSpringDataRepository() throws Exception {
 		String apiPromotionService = Files.readString(
 			Path.of("apis/src/main/java/com/beat/apis/promotion/application/PromotionService.java"));
-		String adminCommandService = Files.readString(
-			Path.of("admin/src/main/java/com/beat/admin/application/service/command/AdminCommandService.java"));
-		String adminQueryService = Files.readString(
-			Path.of("admin/src/main/java/com/beat/admin/application/service/query/AdminQueryService.java"));
+		String adminPromotionCommandService = Files.readString(
+			Path.of("admin/src/main/java/com/beat/admin/promotion/application/service/command/AdminPromotionCommandService.java"));
+		String adminPromotionQueryService = Files.readString(
+			Path.of("admin/src/main/java/com/beat/admin/promotion/application/service/query/AdminPromotionQueryService.java"));
 		String batchSchedulerService = Files.readString(
 			Path.of("batch/src/main/java/com/beat/batch/promotion/application/PromotionMaintenanceService.java"));
 
 		assertTrue(apiPromotionService.contains("import com.beat.domain.promotion.repository.PromotionRepository;"));
-		assertTrue(adminCommandService.contains("import com.beat.domain.promotion.repository.PromotionRepository;"));
-		assertTrue(adminQueryService.contains("import com.beat.domain.promotion.repository.PromotionRepository;"));
+		assertTrue(adminPromotionCommandService.contains("import com.beat.domain.promotion.repository.PromotionRepository;"));
+		assertTrue(adminPromotionQueryService.contains("import com.beat.domain.promotion.repository.PromotionRepository;"));
 		assertTrue(batchSchedulerService.contains("import com.beat.domain.promotion.repository.PromotionRepository;"));
 		assertFalse(apiPromotionService.contains("com.beat.domain.promotion.dao.PromotionRepository"));
-		assertFalse(adminCommandService.contains("com.beat.domain.promotion.dao.PromotionRepository"));
-		assertFalse(adminQueryService.contains("com.beat.domain.promotion.dao.PromotionRepository"));
+		assertFalse(adminPromotionCommandService.contains("com.beat.domain.promotion.dao.PromotionRepository"));
+		assertFalse(adminPromotionQueryService.contains("com.beat.domain.promotion.dao.PromotionRepository"));
 		assertFalse(batchSchedulerService.contains("com.beat.domain.promotion.dao.PromotionRepository"));
 	}
